# Add a benchmark to test restoring from catchpoint performance