## Summary

Before attempting to improve the catchpoint catchup performance, we need a way to systematically test how well we perform. Running this test, produces the following output:
```
goos: darwin
Add a benchmark to test restoring from catchpoint performance.
goarch: amd64
pkg: github.com/algorand/go-algorand/ledger
BenchmarkRestoringFromCatchpointFile/Restore-102400-8         	  102400	     64479 ns/op
--- BENCH: BenchmarkRestoringFromCatchpointFile/Restore-102400-8
    catchupaccessor_test.go:99: Last 64K	4.949638275s
BenchmarkRestoringFromCatchpointFile/Restore-204800-8         	  204800	     86044 ns/op
--- BENCH: BenchmarkRestoringFromCatchpointFile/Restore-204800-8
    catchupaccessor_test.go:99: Last 64K	7.556581895s
BenchmarkRestoringFromCatchpointFile/Restore-409600-8         	  409600	    105683 ns/op
--- BENCH: BenchmarkRestoringFromCatchpointFile/Restore-409600-8
    testingLogger.go:38: time="2020-09-04T12:24:42.944604 -0400" level=warning msg="dbatomic: tx surpassed expected deadline by 11.728647ms" callee="github.com/algorand/go-algorand/ledger.(*CatchpointCatchupAccessorImpl).processStagingBalances.func1" caller="/Users/tsachi/go/src/github.com/algorand/go-algorand/ledger/catchupaccessor.go:294" file=dbutil.go function="github.com/algorand/go-algorand/util/db.(*Accessor).atomic" line=355 readonly=false
    catchupaccessor_test.go:99: Last 64K	9.041332973s
PASS
ok  	github.com/algorand/go-algorand/ledger	68.104s
```

by itself, it's not good or bad. But it's a measurable starting point for performance tuning.